Upgrade `openc_bot` version
In 1f964e9ae6b8f8628e9da660ff7fe0da38214067 `activesupport` was bumped to pick up a security fix. This broke compatibility with `openc_bot` (which had the version pinned to something earlier). `openc_bot` has now relaxed its constraint.

Given `turbotlib` needed to be upgraded as part of this, the version has been pinned exactly to ensure the environment is predicable (which is the whole point of this gem).